requirements.txt
flask>=3.0.0
numpy>=1.24.0
opencv-python>=4.8.0
pyserial>=3.5
ultralytics>=8.3.0
pupil-apriltags>=1.0.4
